Pre-Alpha 0.006e (Tile Problem)
rmTransitionTest created to test transitions

One tile modified on tileset to add addition detail.

Please remember that moving around tiles in the tile image will result
in laid down tiles in existing rooms to be overwritten with the new
image. This will break the existing room layout.

Please append new tiles to the existing image; do not overwrite or move
around existing tiles unless absolutely necessary. (It's okay to do that
with the test tile set, but do not do that with the workin tileset as it
will break rooms.)